# docs/openapi-specification.yaml
openapi: 3.0.3
info:
  title: Thai Auto Insurance API
  version: 1.0.0
  description: |
    Comprehensive REST API for Thai Auto Insurance application supporting policy management, claims processing, and customer service operations.
    
    ## Features
    - Multi-language support (Thai/English)
    - JWT-based authentication
    - Thai-specific validations (National ID, phone numbers, addresses)
    - Real-time payment processing
    - Comprehensive claims management
    - Regulatory compliance (OIC requirements)
    
  contact:
    name: API Support Team
    email: api-support@autoinsurance.co.th
    url: https://docs.autoinsurance.co.th
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://autoinsurance.co.th/terms
servers:
  - url: https://api.autoinsurance.co.th/api/v1
    description: Production server
  - url: https://api-staging.autoinsurance.co.th/api/v1
    description: Staging server
  - url: https://api-dev.autoinsurance.co.th/api/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /claims:
    get:
      tags: [Claims Management]
      summary: Get customer claims
      description: Retrieve all claims for current customer
      operationId: getCustomerClaims
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [REPORTED, UNDER_INVESTIGATION, APPROVED, SETTLED, REJECTED]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Claims retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimListResponse'

    post:
      tags: [Claims Management]
      summary: Report new claim
      description: Report insurance claim with supporting documents
      operationId: reportClaim
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReportClaimRequest'
      responses:
        '201':
          description: Claim reported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportClaimResponse'
  /claims/{claimId}:
    get:
      tags: [Claims Management]
      summary: Get claim details
      description: Retrieve detailed claim information
      operationId: getClaimDetails
      parameters:
        - name: claimId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Claim details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimDetailsResponse'
  /claims/{claimId}/status:
    put:
      tags: [Claims Management]
      summary: Update claim status
      description: Update claim processing status (Admin only)
      operationId: updateClaimStatus
      parameters:
        - name: claimId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClaimStatusRequest'
      responses:
        '200':
          description: Claim status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateClaimStatusResponse'
...
  /admin/claims:
    get:
      tags: [Admin Operations]
      summary: Get all claims (Admin)
      description: Retrieve paginated list of all claims
      operationId: getAllClaims
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [REPORTED, UNDER_INVESTIGATION, APPROVED, SETTLED, REJECTED]
      responses:
        '200':
          description: Claims retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminClaimListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    ClaimListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            claims:
              type: array
              items:
                $ref: '#/components/schemas/ClaimSummary'
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
    ClaimSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        claimNumber:
          type: string
        policyNumber:
          type: string
        incidentType:
          type: string
        incidentDate:
          type: string
          format: date
        status:
          type: string
          enum: [REPORTED, UNDER_INVESTIGATION, APPROVED, SETTLED, REJECTED]
        estimatedAmount:
          type: number
          format: float
        reportedAt:
          type: string
          format: date-time
    ReportClaimRequest:
      type: object
      required:
        - policyId
        - claimType
        - incidentDate
        - description
      properties:
        policyId:
          type: string
          format: uuid
        claimType:
          type: string
          enum: [ACCIDENT, THEFT, FIRE, FLOOD, VANDALISM, NATURAL_DISASTER]
        incidentDate:
          type: string
          format: date-time
        incidentLocation:
          $ref: '#/components/schemas/IncidentLocation'
        description:
          type: string
          minLength: 10
          maxLength: 1000
        policeReportNumber:
          type: string
        otherPartyInfo:
          $ref: '#/components/schemas/OtherPartyInfo'
        damageDescription:
          type: string
        injuryDescription:
          type: string
        photos:
          type: array
          items:
            type: string
            format: binary
        videos:
          type: array
          items:
            type: string
            format: binary
        documents:
          type: array
          items:
            type: string
            format: binary
    IncidentLocation:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
        landmark:
          type: string
    OtherPartyInfo:
      type: object
      properties:
        driverName:
          type: string
        licenseNumber:
          type: string
        vehicleRegistration:
          type: string
        insuranceCompany:
          type: string
    ReportClaimResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            claimId:
              type: string
              format: uuid
            claimNumber:
              type: string
              example: "CLM-2024-123456"
            status:
              type: string
              example: "REPORTED"
            reportedAt:
              type: string
              format: date-time
            estimatedProcessingTime:
              type: string
              example: "7-10 business days"
            assignedAdjuster:
              $ref: '#/components/schemas/AssignedAdjuster'
            nextSteps:
              type: array
              items:
                type: string
            trackingUrl:
              type: string
              format: uri
            message:
              type: string
              example: "Claim reported successfully. You will receive updates via SMS and email."
        timestamp:
          type: string
          format: date-time
    AssignedAdjuster:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
    ClaimDetailsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/ClaimDetails'
        timestamp:
          type: string
          format: date-time
    ClaimDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
        claimNumber:
          type: string
        policy:
          type: object
          properties:
            policyNumber:
              type: string
            coverageType:
              type: string
        incident:
          $ref: '#/components/schemas/IncidentDetails'
        damages:
          $ref: '#/components/schemas/DamageAssessment'
        assessment:
          $ref: '#/components/schemas/ClaimAssessment'
        repairShop:
          $ref: '#/components/schemas/RepairShopInfo'
        status:
          type: string
        settlement:
          $ref: '#/components/schemas/SettlementInfo'
    IncidentDetails:
      type: object
      properties:
        type:
          type: string
        date:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/IncidentLocation'
        description:
          type: string
        policeReportNumber:
          type: string
    DamageAssessment:
      type: object
      properties:
        vehicleDamage:
          type: object
          properties:
            estimatedCost:
              type: number
              format: float
            description:
              type: string
            photos:
              type: array
              items:
                type: string
                format: uri
        medicalExpenses:
          type: object
          properties:
            estimatedCost:
              type: number
              format: float
            description:
              type: string
            hospitalReports:
              type: array
              items:
                type: string
                format: uri
    ClaimAssessment:
      type: object
      properties:
        surveyor:
          $ref: '#/components/schemas/Surveyor'
        assessmentDate:
          type: string
          format: date-time
        estimatedRepairCost:
          type: number
          format: float
        deductible:
          type: number
          format: float
        approvedAmount:
          type: number
          format: float
    Surveyor:
      type: object
      properties:
        name:
          type: string
        licenseNumber:
          type: string
        phone:
          type: string
    RepairShopInfo:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        estimatedCompletionDate:
          type: string
          format: date
    SettlementInfo:
      type: object
      properties:
        approvedAmount:
          type: number
          format: float
        paymentMethod:
          type: string
          enum: [DIRECT_REPAIR, CASH_SETTLEMENT, BANK_TRANSFER]
        paymentDate:
          type: string
          format: date
    UpdateClaimStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [UNDER_INVESTIGATION, APPROVED, SETTLED, REJECTED]
        notes:
          type: string
        assignedTo:
          type: string
    UpdateClaimStatusResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            claimId:
              type: string
              format: uuid
            previousStatus:
              type: string
            newStatus:
              type: string
            updatedAt:
              type: string
              format: date-time
            updatedBy:
              type: string
            message:
              type: string
        timestamp:
          type: string
          format: date-time
...
    AdminClaimListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            claims:
              type: array
              items:
                $ref: '#/components/schemas/AdminClaimSummary'
            pagination:
              $ref: '#/components/schemas/Pagination'
        timestamp:
          type: string
          format: date-time
    AdminClaimSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        claimNumber:
          type: string
        customerName:
          type: string
        policyNumber:
          type: string
        incidentType:
          type: string
        claimAmount:
          type: number
          format: float
        status:
          type: string
        assignedTo:
          type: string
        reportedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...